refactored Timer and created ButtonTimerStart component
-useTime custom hook now handles countdown and time state in Timer component
-logic for the button, which changes whether timer runs or not, now transferred to ButtonTimerStart component
